Added a Cranium Class
Made a few functions for the game,
dice roller function
function to select a random player
gui to pull in data from the user on the number of players and all the
error handling with it.
blobzx is an import to use the sandbox to make a window of standard size
that has a bunch of useful functions for little game gui's

the images are the board which will be used for the background of the
sandbox.

the card example is for the artsy people to have as a layout