Simpler version of Card.java
joanhquez:

This is just a simple Card class I created based on the Card.java file on here. I used it to see how the cards I made were going to look. Courtney, since you're doing the card display stuff, feel free to use any of the code in this or if you need any help just let me know.